requirements.txt
anthropic==0.74.1
openpyxl==3.1.5
pdf2image==1.17.0
Pillow==12.0.0
python-dotenv==1.2.1
